@@ -27,66 +27,67 @@
   },
   "homepage": "https://owncloud.github.io/integration/file_picker",
   "scripts": {
-    "serve": "vue-cli-service serve",
-    "build:wc": "vue-cli-service build --target wc --name file-picker --dest ./dist/wc",
-    "build:lib": "vue-cli-service build --target lib --name file-picker --dest ./dist/lib",
-    "test:unit": "TEST_MODE=unit vue-cli-service test:unit",
-    "test:integration": "TEST_MODE=integration jest --config ./jest.config.js --env jest-environment-jsdom-sixteen",
-    "lint": "vue-cli-service lint",
-    "build": "yarn build:wc && yarn build:lib"
-  },
-  "dependencies": {
-    "core-js": "^3.11.1",
-    "filesize": "^6.3.0",
-    "focus-trap": "^6.5.0",
-    "focus-trap-vue": "^1.1.1",
-    "lodash": "^4.17.21",
-    "lodash-es": "^4.17.21",
-    "luxon": "^1.26.0",
-    "oidc-client": "^1.11.5",
-    "owncloud-design-system": "^7.2.0",
-    "owncloud-sdk": "^1.0.0-2296",
-    "postcss-import": "^12.0.1",
-    "postcss-url": "^9.0.0",    
-    "tippy.js": "^6.3.1",
-    "uikit": "3.5.16",
-    "vue-avatar": "^2.3.3",
-    "vue-datetime": "^1.0.0-beta.14",
-    "vue-gettext": "^2.1.12",
-    "vue-inline-svg": "^2.0.0",
-    "vue-select": "^3.11.2",
-    "webfontloader": "^1.6.28",
-    "weekstart": "^1.0.0"
+    "dev": "vite",
+    "build:lib": "vue-tsc --noEmit && vite build --mode library",
+    "build:wc": "vue-tsc --noEmit && vite build --mode web-component",
+    "build": "BABEL_ENV=production rm -rf ./dist && pnpm build:lib && pnpm build:wc",
+    "preview": "cp -r ./examples/ocis/ ./dist/wc && vite preview --mode web-component",
+    "lint": "eslint",
+    "test:unit": "BABEL_ENV=test TEST_MODE=unit jest --config ./jest.config.cjs",
+    "test:integration": "BABEL_ENV=test TEST_MODE=integration jest --config ./jest.config.cjs",
+    "postinstall": "bash ./scripts/install-ods.sh"
   },
   "devDependencies": {
-    "@babel/core": "^7.13.16",
-    "@babel/eslint-parser": "^7.13.14",
+    "@babel/core": "^7.21.4",
+    "@babel/helper-plugin-utils": "^7.20.2",
+    "@babel/plugin-transform-modules-commonjs": "^7.21.2",
     "@babel/polyfill": "^7.12.1",
-    "@testing-library/jest-dom": "^5.12.0",
-    "@testing-library/vue": "^5.6.2",
-    "@vue/cli-plugin-babel": "^4.5.12",
-    "@vue/cli-plugin-eslint": "^4.5.12",
-    "@vue/cli-plugin-unit-jest": "^4.5.12",
-    "@vue/cli-service": "^4.5.12",
-    "@vue/eslint-config-prettier": "^6.0.0",
-    "@vue/test-utils": "^1.1.4",
-    "babel-core": "^7.0.0-bridge.0",
-    "babel-jest": "^26.6.3",
-    "easygettext": "^2.17.0",
-    "eslint": "^7.25.0",
-    "eslint-plugin-prettier": "^3.4.0",
-    "eslint-plugin-vue": "^7.9.0",
-    "jest-environment-jsdom-sixteen": "^2.0.0",
+    "@babel/preset-env": "^7.21.4",
+    "@babel/register": "^7.21.0",
+    "@ownclouders/babel-preset": "^0.0.1",
+    "@ownclouders/eslint-config": "^0.0.1",
+    "@ownclouders/prettier-config": "^0.0.1",
+    "@ownclouders/tsconfig": "^0.0.4",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/vue": "^5.9.0",
+    "@types/node": "^18.15.11",
+    "@vitejs/plugin-basic-ssl": "^1.0.1",
+    "@vitejs/plugin-vue2": "^2.2.0",
+    "@vue/babel-preset-app": "^5.0.8",
+    "@vue/test-utils": "^1.3.4",
+    "@vue/vue2-jest": "^29.2.3",
+    "babel-jest": "^29.5.0",
+    "buffer": "^6.0.3",
+    "core-js": "^3.30.0",
+    "dotenv": "^16.0.3",
+    "eslint": "^8.37.0",
+    "jest": "^29.5.0",
+    "jest-environment-jsdom": "^29.5.0",
     "jest-fetch-mock": "^3.0.3",
-    "jsdom": "^16.5.3",
-    "prettier": "^2.2.1",
-    "sass": "^1.32.12",
-    "sass-loader": "10",
-    "vue": "^2.6.12",
-    "vue-jest": "^3.0.7",
-    "vue-loader": "^15.9.6",
-    "vue-template-compiler": "^2.6.12",
-    "webpack": "^4.0.0"
+    "jest-serializer-vue": "^3.1.0",
+    "jsdom": "^21.1.1",
+    "lodash": "^4.17.21",
+    "owncloud-design-system": "github:owncloud/owncloud-design-system#release-15.0.0",
+    "owncloud-sdk": "3.1.0-alpha.4",
+    "sass": "^1.60.0",
+    "stream-browserify": "^3.0.0",
+    "typescript": "^4.9.3",
+    "vite": "3.2.5",
+    "vue": "2.7.13",
+    "vue-template-compiler": "2.7.13",
+    "vue-tsc": "^1.2.0"
+  },
+  "dependencies": {
+    "@vue/web-component-wrapper": "^1.3.0",
+    "deepmerge": "^4.3.1",
+    "lodash-es": "^4.17.21",
+    "luxon": "^3.3.0",
+    "oidc-client": "^1.11.5",
+    "vue-gettext": "^2.1.12"
+  },
+  "pnpm": {
+    "patchedDependencies": {
+      "@vue/web-component-wrapper@1.3.0": "patches/@vue__web-component-wrapper@1.3.0.patch"
+    }
   }
 }
-
